# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SU Ples - Api Documentation",
        "description": "Api Documentation for UTB Representative Ball",
        "contact": {
            "name": "Sedlar David",
            "email": "sedlar@sutb.cz"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/api/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get list of reservations",
                "description": "Get list of reservations",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Creates a reservation",
                "operationId": "01e3d21dc330a31de025203bc142ffa3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "tel": {
                                        "type": "integer"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "stand": {
                                        "type": "integer"
                                    },
                                    "seats": {
                                        "type": "integer"
                                    },
                                    "": {
                                        "description": "List of Seat ids",
                                        "type": "array",
                                        "items": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "David Sedlar",
                                    "email": "sedlar@sutb.cz",
                                    "tel": 555222555,
                                    "note": "Popici ples, chci celej stul...",
                                    "stand": 3,
                                    "seats": [
                                        2,
                                        3,
                                        5
                                    ]
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/seats": {
            "get": {
                "tags": [
                    "Seats"
                ],
                "summary": "Get list of seats",
                "description": "Get list of seats",
                "operationId": "09a0c40bbde096321060e7db24676668",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    }
}